mindful-feed/backend/requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.0
supabase>=2.4.0
pyyaml>=6.0.1
